ax2-chatbot-back/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-community
pyyaml
